refac: add routing when tapping button
