+++ college_flow_app/lib/config/routes/flow_router.dart
@@ -1,12 +1,14 @@
 import 'package:college_flow_app/config/routes/flow_routes.dart';
 import 'package:college_flow_app/presentation/login/login_pages.dart';
+import 'package:college_flow_app/presentation/register/admin/register_adm.dart';
+import 'package:college_flow_app/presentation/register/teacher/register_teacher.dart';
 import 'package:college_flow_app/presentation/walkthrough/walkthrough_page.dart';
 import 'package:college_flow_app/utils/helpers/navigator_helper.dart';
 import 'package:flutter/material.dart' hide Router;
 
 import '../../presentation/register/register_geral.dart';
 import '../../presentation/register/register_page.dart';
-import '../../presentation/register/register_student.dart';
+import '../../presentation/register/student/register_student.dart';
 
 abstract class FlowRouter {
   static Route routeFromSettings(RouteSettings settings) {
@@ -18,6 +20,12 @@ abstract class FlowRouter {
         break;
       case FlowRoutes.registerStudent:
         page = const RegisterStudent();
+        break;
+      case FlowRoutes.registerAdmin:
+        page = const RegisterAdministrator();
+        break;
+      case FlowRoutes.registerProfessor:
+        page = const RegisterTeacher();
         break;
       case FlowRoutes.registerPage:
         page = const RegisterPage();
