+++ college_flow_app/lib/config/routes/flow_routes.dart
@@ -1,7 +1,9 @@
 abstract class FlowRoutes {
   static const mainPage = '/';
   static const loginPage = '/LoginPage';
-  static const registerPage = '/RegisterPage';
-  static const registerGeral = '/RegisterGeral';
-  static const registerStudent = '/RegisterStudent';
+  static const registerPage = '/registerPage';
+  static const registerGeral = '/registerGeral';
+  static const registerStudent = '/registerStudent';
+  static const registerProfessor = '/registerProfessor';
+  static const registerAdmin = '/registerAdmin';
 }
