+++ college_flow_app/lib/presentation/register/admin/register_adm.dart
@@ -1,8 +1,8 @@
 import 'dart:ffi';
-
 import 'package:college_flow_app/config/design_system/data/colors/colors.dart';
 import 'package:college_flow_app/config/design_system/data/spacing/spacing.dart';
-import 'package:college_flow_app/presentation/register/register_adm.dart';
+import 'package:college_flow_app/presentation/register/teacher/register_teacher.dart';
+
 import 'package:college_flow_app/presentation/widgets/buttons/flow_button.dart';
 import 'package:college_flow_app/presentation/widgets/gap.dart';
 import 'package:college_flow_app/presentation/widgets/textfield/flow_text_field.dart';
@@ -10,14 +10,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/src/foundation/key.dart';
 import 'package:flutter/src/widgets/framework.dart';
 
-class RegisterStudent extends StatefulWidget {
-  const RegisterStudent({Key? key}) : super(key: key);
+class RegisterAdministrator extends StatefulWidget {
+  const RegisterAdministrator({Key? key}) : super(key: key);
 
   @override
-  State<RegisterStudent> createState() => _RegisterStudentState();
+  State<RegisterAdministrator> createState() => _RegisterAdministratorState();
 }
 
-class _RegisterStudentState extends State<RegisterStudent> {
+class _RegisterAdministratorState extends State<RegisterAdministrator> {
   final _pageViewController = PageController();
 
   @override
@@ -56,40 +56,25 @@ class _RegisterStudentState extends State<RegisterStudent> {
             ),
             const VSpacer.xxs(),
             Text(
-              'Matrícula',
+              'Insira Abaixo o seu Código de Administrador',
+              textAlign: TextAlign.center,
+              style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+                    color: textColor,
+                  ),
+            ),
+            const VSpacer.xs(),
+            Text(
+              'Código',
               style: Theme.of(context).textTheme.titleMedium?.copyWith(
                     color: textColor,
                   ),
             ),
             const VSpacer.nano(),
             const FlowTextField(
-              label: "E-mail",
-              placeholder: 'Digite sua Matrícula',
+              label: "Código de Administrador",
+              placeholder: 'Digite seu código',
             ),
             const VSpacer.xxxs(),
-            Text(
-              'Curso',
-              style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                    color: textColor,
-                  ),
-            ),
-            const VSpacer.nano(),
-            const FlowTextField(
-              label: "Curso",
-              placeholder: 'selecione seu curso',
-            ),
-            const VSpacer.xxxs(),
-            Text(
-              'Semestre',
-              style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                    color: textColor,
-                  ),
-            ),
-            const VSpacer.nano(),
-            const FlowTextField(
-              label: "Semestre",
-              placeholder: 'Digite seu semestre',
-            ),
             const VSpacer.xxs(),
             FlowButton(
               label: "Continue",
@@ -99,6 +84,14 @@ class _RegisterStudentState extends State<RegisterStudent> {
           ],
         ),
       ),
+    );
+  }
+
+  void togglePage(int page) {
+    _pageViewController.animateToPage(
+      page,
+      duration: const Duration(milliseconds: 200),
+      curve: Curves.easeIn,
     );
   }
 }
